refactor: general structure
