+++ controllers/music/features/players/getMessages.js
@@ -1,13 +1,16 @@
-const filterLinks = require('./filterLinks')
+const formatLinks = require('./formatLinks')
 const getLinks = require('./getLinks')
-const MessageHelper = require("../../message/MessageHelper")
-const MESSAGES = require("../../message/messages.enum")
+const getSpecificLinks = require('./getSpecificLinks')
+
+const MessageHelper = require('../../../message/MessageHelper')
+const MESSAGES = require('../../../message/messages.enum')
 
 async function getMessages(music) {
     let content = []
     
-    const links = await getLinks(music)
-    let linksMessage = await filterLinks(links)
+    const response = await getLinks(music.spotify.link)
+    const links = getSpecificLinks(response.linksByPlatform)
+    let linksMessage = await formatLinks(links)
 
     if(linksMessage) {
         linksMessage = '*Players de Música:*\n\n' + linksMessage
