+++ controllers/music/features/tradutions/getMessages.js
@@ -1,7 +1,7 @@
 const axios = require("axios")
 const dotenv = require('dotenv')
-const MessageHelper = require("../../message/MessageHelper")
-const MESSAGES = require("../../message/messages.enum")
+const MessageHelper = require("../../../message/MessageHelper")
+const MESSAGES = require("../../../message/messages.enum")
 
 async function getMessages(music) {
     let content = []
