--- controllers/music/infos/Infos.js
@@ -1,13 +1,0 @@
-const getMessages = require("./getMessages")
-
-class Infos {
-    constructor(music) {
-        this.music = music
-    }
-
-    async getMessages() {
-        return await getMessages(this.music);
-    }
-}
-
-module.exports = Infos
