--- controllers/music/players/Players.js
@@ -1,13 +1,0 @@
-const getMessages = require('./getMessages');
-
-class Players {
-    constructor(music) {
-        this.music = music
-    }
-
-    async getMessages() {
-        return await getMessages(this.music)
-    }
-}
-
-module.exports = Players
