+++ index.js
@@ -1,0 +1,3 @@
+const Webhook = require('./controllers/webhook/Webhook')
+const webhook = new Webhook('whatsapp')
+webhook.boot()
