--- src/index.js
@@ -1,3 +1,0 @@
-const Webhook = require('../controllers/webhook/Webhook')
-const webhook = new Webhook('whatsapp')
-webhook.boot()
